fix: bug fix

+++ firebaseConfig.tsx
@@ -1,14 +1,12 @@
 import { initializeApp, FirebaseApp } from "firebase/app";
 import { getFirestore, Firestore, collection, addDoc, getDocs, getDoc, deleteDoc, doc } from "firebase/firestore";
 import { getAuth, Auth, signInWithEmailAndPassword } from "firebase/auth";
-import { getStorage } from "firebase/storage"; // Correct import
-import { getAnalytics, Analytics } from "firebase/analytics";
+import { getStorage } from "firebase/storage";
 
 let app: FirebaseApp | undefined;
 let db: Firestore | undefined;
 let auth: Auth | undefined;
-let storage: ReturnType<typeof getStorage> | undefined; // Use ReturnType to type storage
-let analytics: Analytics | undefined;
+let storage: ReturnType<typeof getStorage> | undefined;
 
 const firebaseConfig = {
   apiKey: "AIzaSyAuny5xpmnG3WJ66hPi3RtunJVFXPm1AcM",
@@ -26,7 +24,11 @@ if (typeof window !== "undefined" && !app) {
   db = getFirestore(app);
   auth = getAuth(app);
   storage = getStorage(app); // Initialize Firebase Storage
-  analytics = getAnalytics(app);
+}
+
+// Ensure that auth and other variables are defined before exporting
+if (!auth || !db || !storage) {
+  throw new Error("Firebase services are not properly initialized");
 }
 
 export { db, collection, addDoc, getDocs, getDoc, deleteDoc, doc, auth, signInWithEmailAndPassword, storage, firebaseConfig };
